fix models

+++ src/modules/teachers/services/PositionService.js
@@ -3,33 +3,41 @@ import NotFound from '../../../classes/errors/4xx/notFound'
 
 class PositionService {
 
-    async list() {
-        return await PositionRepository.readAll()
-    }
-
     async create(position) {
         return await PositionRepository.create(position)
     }
 
-    /*  ????  */
     async findById(id) {
+
         let position = await PositionRepository.readById(id)
+
         if (!position) {
-            throw new NotFound(`${objectName} not found`)
+            throw new NotFound(`Position not found`)
         }
+
         return position
     }
 
     async update(id, position) {
-        let nPosition = PositionRepository.readById(id)
-        if(!nPosition) {            
-            throw new NotFound(`${objectName} not found`)
+
+        let oldPosition = PositionRepository.readById(id)
+
+        if(!oldPosition) {
+            throw new NotFound(`Position not found`)
         }
+
         return await PositionRepository.update(id, position)
     }
 
     async destroy(id) {
-        return await PositionRepository.destroy(id)
+
+        let oldPosition = PositionRepository.readById(id)
+
+        if(!oldPosition) {
+            throw new NotFound(`Position not found`)
+        }
+
+        await PositionRepository.destroy(id)
     }
 }
